pyproject.toml
[project]
name = "cv-jd-orchestrator"
version = "0.1.0"
description = "Agentic CV-vs-JD alignment workflow with FastAPI, LangGraph, Google Workspace, Gmail, and YouTube integrations."
readme = "README.md"
keywords = [
    "fastapi",
    "langgraph",
    "llm",
    "google-workspace",
    "gmail",
    "openai",
    "cv",
    "job-description",
]
authors = [{name = "Shiva Sharma"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "langgraph>=0.0.43",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "email-validator>=2.1",
    "openai>=1.30",
    "google-api-python-client>=2.120",
    "google-auth>=2.32",
    "google-genai>=0.5.0",
    "httpx>=0.27",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "redis>=5.0",
    "itsdangerous>=2.2",
    "google-auth-oauthlib>=1.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "anyio>=4.3"
]

[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app", "src/orchestrator", "src/services"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["src/tests"]
asyncio_mode = "auto"
